Extract useOpenResource helper to open a source location generally
