+++ packages/react-devtools-shared/src/devtools/views/Components/InspectedElementSourcePanel.js
@@ -8,7 +8,6 @@
  */
 
 import * as React from 'react';
-import {useCallback, useContext} from 'react';
 import {copy} from 'clipboard-js';
 import {toNormalUrl} from 'jsc-safe-url';
 
@@ -17,7 +16,7 @@ import ButtonIcon from '../ButtonIcon';
 import Skeleton from './Skeleton';
 import {withPermissionsCheck} from 'react-devtools-shared/src/frontend/utils/withPermissionsCheck';
 
-import ViewElementSourceContext from './ViewElementSourceContext';
+import useOpenResource from '../useOpenResource';
 
 import type {ReactFunctionLocation} from 'shared/ReactTypes';
 import styles from './InspectedElementSourcePanel.css';
@@ -91,23 +90,10 @@ function CopySourceButton({source, symbolicatedSourcePromise}: Props) {
 function FormattedSourceString({source, symbolicatedSourcePromise}: Props) {
   const symbolicatedSource = React.use(symbolicatedSourcePromise);
 
-  const {canViewElementSourceFunction, viewElementSourceFunction} = useContext(
-    ViewElementSourceContext,
+  const [linkIsEnabled, viewSource] = useOpenResource(
+    source,
+    symbolicatedSource,
   );
-
-  // In some cases (e.g. FB internal usage) the standalone shell might not be able to view the source.
-  // To detect this case, we defer to an injected helper function (if present).
-  const linkIsEnabled =
-    viewElementSourceFunction != null &&
-    source != null &&
-    (canViewElementSourceFunction == null ||
-      canViewElementSourceFunction(source, symbolicatedSource));
-
-  const viewSource = useCallback(() => {
-    if (viewElementSourceFunction != null && source != null) {
-      viewElementSourceFunction(source, symbolicatedSource);
-    }
-  }, [source, symbolicatedSource]);
 
   const [, sourceURL, line] =
     symbolicatedSource == null ? source : symbolicatedSource;
